+++ src/cpp/tests/test_hydra.cpp
@@ -618,12 +618,126 @@ BOOST_AUTO_TEST_CASE( test_hydra_getSubConfiguration ){
     hydra::hydraBase hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
                             previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
 
-    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getSubConfiguration( 0, 3 ), *hydra.getDeformationGradient( ) ) );
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getSubConfiguration( 0, 4 ), *hydra.getDeformationGradient( ) ) );
 
     floatVector answer1 = { 2.24332648, 1.48246714, 2.02801682,
                             1.50380989, 2.98203598, 2.08079721,
                             1.58939152, 1.2551092 , 2.38201794 };
 
-    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getSubConfiguration( 1, 2 ), answer1 ) );
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getSubConfiguration( 1, 3 ), answer1 ) );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getPrecedingConfiguration ){
+
+    floatType time = 1.1;
+
+    floatType deltaTime = 2.2;
+
+    floatType temperature = 5.3;
+
+    floatType previousTemperature = 23.4;
+
+    floatVector deformationGradient = { 0.39293837, -0.42772133, -0.54629709,
+                                        0.10262954,  0.43893794, -0.15378708,
+                                        0.9615284 ,  0.36965948, -0.0381362 };
+
+    floatVector previousDeformationGradient = { -0.21576496, -0.31364397,  0.45809941,
+                                                -0.12285551, -0.88064421, -0.20391149,
+                                                 0.47599081, -0.63501654, -0.64909649 };
+
+    floatVector previousStateVariables = { 0.53155137, 0.53182759, 0.63440096, 0.84943179, 0.72445532,
+                                           0.61102351, 0.72244338, 0.32295891, 0.36178866, 0.22826323,
+                                           0.29371405, 0.63097612, 0.09210494, 0.43370117, 0.43086276,
+                                           0.4936851 , 0.42583029, 0.31226122, 0.42635131, 0.89338916,
+                                           0.94416002, 0.50183668, 0.62395295, 0.1156184 , 0.31728548,
+                                           0.41482621, 0.86630916, 0.25045537, 0.48303426, 0.98555979,
+                                           0.51948512, 0.61289453, 0.12062867, 0.8263408 , 0.60306013,
+                                           0.54506801, 0.34276383, 0.30412079 }; 
+
+    floatVector parameters = { 1, 2, 3, 4, 5 };
+
+    unsigned int numConfigurations = 4;
+
+    unsigned int numNonLinearSolveStateVariables = 5;
+
+    unsigned int dimension = 3;
+
+    hydra::hydraBase hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                            previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getPrecedingConfiguration( 4 ), *hydra.getDeformationGradient( ) ) );
+
+    floatVector answer1 = { 0.73947165, -0.24328161, -0.68904986,
+                            0.04049558,  0.25403825, -0.1748363 ,
+                            0.97015752, -0.04452644, -0.77301275 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getPrecedingConfiguration( 2 ), answer1 ) );
+
+    floatVector answer2 = { 0.54568475, -0.42501821, -0.54244544,
+                           -0.01317666,  0.30165847, -0.09442353,
+                            0.80588282, -0.10806097, -0.42143322 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getPrecedingConfiguration( 3 ), answer2 ) );
+
+    floatVector answer3 = { 1, 0, 0, 0, 1, 0, 0, 0, 1 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getPrecedingConfiguration( 0 ), answer3 ) );
+
+}
+
+BOOST_AUTO_TEST_CASE( test_hydra_getFollowingConfiguration ){
+
+    floatType time = 1.1;
+
+    floatType deltaTime = 2.2;
+
+    floatType temperature = 5.3;
+
+    floatType previousTemperature = 23.4;
+
+    floatVector deformationGradient = { 0.39293837, -0.42772133, -0.54629709,
+                                        0.10262954,  0.43893794, -0.15378708,
+                                        0.9615284 ,  0.36965948, -0.0381362 };
+
+    floatVector previousDeformationGradient = { -0.21576496, -0.31364397,  0.45809941,
+                                                -0.12285551, -0.88064421, -0.20391149,
+                                                 0.47599081, -0.63501654, -0.64909649 };
+
+    floatVector previousStateVariables = { 0.53155137, 0.53182759, 0.63440096, 0.84943179, 0.72445532,
+                                           0.61102351, 0.72244338, 0.32295891, 0.36178866, 0.22826323,
+                                           0.29371405, 0.63097612, 0.09210494, 0.43370117, 0.43086276,
+                                           0.4936851 , 0.42583029, 0.31226122, 0.42635131, 0.89338916,
+                                           0.94416002, 0.50183668, 0.62395295, 0.1156184 , 0.31728548,
+                                           0.41482621, 0.86630916, 0.25045537, 0.48303426, 0.98555979,
+                                           0.51948512, 0.61289453, 0.12062867, 0.8263408 , 0.60306013,
+                                           0.54506801, 0.34276383, 0.30412079 }; 
+
+    floatVector parameters = { 1, 2, 3, 4, 5 };
+
+    unsigned int numConfigurations = 4;
+
+    unsigned int numNonLinearSolveStateVariables = 5;
+
+    unsigned int dimension = 3;
+
+    hydra::hydraBase hydra( time, deltaTime, temperature, previousTemperature, deformationGradient, previousDeformationGradient,
+                            previousStateVariables, parameters, numConfigurations, numNonLinearSolveStateVariables, dimension );
+
+    floatVector answer1 = { 2.09953091, 1.83604029, 2.3712323 ,
+                            0.98756433, 2.58928197, 1.05684715,
+                            1.33422708, 1.67694162, 2.96443669 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getFollowingConfiguration( 1 ), answer1 ) );
+
+    floatVector answer2 = { 1.42635131, 0.89338916, 0.94416002,
+                            0.50183668, 1.62395295, 0.1156184 ,
+                            0.31728548, 0.41482621, 1.86630916 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getFollowingConfiguration( 2 ), answer2 ) );
+
+    floatVector answer3 = { 1, 0, 0, 0, 1, 0, 0, 0, 1 };
+
+    BOOST_CHECK( vectorTools::fuzzyEquals( hydra.getFollowingConfiguration( 3 ), answer3 ) );
 
 }